Fail validation for very long email addresses
Identity can’t handle addresses longer than 100 characters in some of
its endpoints, and returns a 500. Since we don’t want alerts in this
case, but we do want alerts on 500s, this change fails validation to
avoid calling Identity and return a 400.

There is corresponding client-side validation as well, in
https://github.com/guardian/support-dotcom-components/pull/846

One problem with this is: what if Identity fix this in the future, and
accept emails longer than 100 characters at creation? Then someone could
make an account but not be able to sign up for reminders. I’m not sure
what to do about this, but I’ve reached out to Identity to start the
conversation.